ویرایش گروهی درست شد و قابل استفاده است.
